gadget,gadget/install: move fs deployment to install subpackage
MakeFilesystems, DeployContent and MountFilesystem to the gadget/install
subpackage.
