+++ gadget/install/export_test.go
@@ -17,13 +17,28 @@
  *
  */
 
-package gadget
+package install
 
-import (
-	"syscall"
-)
+func MockDeployMountpoint(new string) (restore func()) {
+	old := deployMountpoint
+	deployMountpoint = new
+	return func() {
+		deployMountpoint = old
+	}
+}
 
-var (
-	sysMount   = syscall.Mount
-	sysUnmount = syscall.Unmount
-)
+func MockSysMount(f func(source, target, fstype string, flags uintptr, data string) error) (restore func()) {
+	old := sysMount
+	sysMount = f
+	return func() {
+		sysMount = old
+	}
+}
+
+func MockSysUnmount(f func(target string, flags int) error) (restore func()) {
+	old := sysUnmount
+	sysUnmount = f
+	return func() {
+		sysUnmount = old
+	}
+}
